@@ -1,6 +1,6 @@
 name: flutter_triple
 description: Implementation of the Segmented State Pattern (SSP) nicknamed Triple.
-version: 1.2.8-dev.1
+version: 3.0.0
 
 repository: https://github.com/Flutterando/triple_pattern/tree/master/stores/flutter_triple
 issue_tracker: https://github.com/Flutterando/triple_pattern/issues
@@ -8,57 +8,22 @@ homepage: https://triple.flutterando.com.br/
 documentation: https://triple.flutterando.com.br/
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <4.0.0"
 
 dependencies:
-  triple: ">=1.5.0+1 <2.0.0"
-  rx_notifier: ">=1.1.0 <2.0.0"
+  triple: ">=2.0.1 <3.0.0"
   collection: ^1.15.0
   flutter:
     sdk: flutter
+  asp: ">=1.3.0 <2.0.0"
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutterando_analysis: ^0.0.1
+  flutterando_analysis: 0.0.2
+  equatable: ^2.0.5
+  
 
-# dependency_overrides:
-#   triple: 
-#     path: ../triple
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
 
-  # To add assets to your package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # To add custom fonts to your package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/custom-fonts/#from-packages
