@@ -1,10 +1,10 @@
 name: triple
 description: Abstraction for the Segmented State Pattern. Use to create Triple Store. 
-version: 1.5.0+1
+version: 2.1.0
 homepage: https://github.com/Flutterando/triple_pattern
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <4.0.0'
 
 dependencies:
   meta: ">=1.3.0 <2.0.0"
@@ -12,4 +12,4 @@ dependencies:
 
 dev_dependencies:
   test: ^1.16.5
-  flutterando_analysis: ^0.0.1
+  flutterando_analysis: ^0.0.2
