@@ -1,7 +1,7 @@
 import 'package:flutter_triple/flutter_triple.dart';
 
-class Counter extends StreamStore<Exception, int> with MementoMixin {
-  Counter() : super(0);
+class CounterStore extends Store<int> with MementoMixin {
+  CounterStore() : super(0);
 
   Future<void> increment() async {
     setLoading(true);
