@@ -59,7 +59,7 @@ void main() {
 }
 
 // ignore: must_be_immutable
-class Counter extends TestImplements<Exception, int> {
+class Counter extends TestImplements<int> {
   Counter(List<bool> list) : super(0, list);
 
   FutureOr<void> increment() => execute(
@@ -75,20 +75,19 @@ class Counter extends TestImplements<Exception, int> {
 }
 
 // ignore: must_be_immutable
-abstract class TestImplements<Error extends Object, State extends Object>
-    extends Store<Error, State> {
+abstract class TestImplements<State> extends BaseStore<State> {
   final List<bool> list;
 
   TestImplements(State initialState, this.list) : super(initialState);
 
-  late Triple<Error, State> propagated = triple;
+  late Triple<State> propagated = triple;
 
   @override
   Future destroy() async {}
 
   @override
   void setLoading(bool newloading, {bool force = false}) {
-    super.setLoading(newloading);
+    super.setLoading(newloading, force: force);
     list.add(newloading);
   }
 
@@ -103,7 +102,7 @@ abstract class TestImplements<Error extends Object, State extends Object>
 
   @protected
   @override
-  void propagate(Triple<Error, State> triple) {
+  void propagate(Triple<State> triple) {
     super.propagate(triple);
     propagated = triple;
   }
